requirements.txt
emoji==0.6.0
pytz==2020.5
requests==2.20.0
python_telegram_bot==13.1
